remove memory fixes

+++ src/xray/xrCore/xrMemory_pso_Fill32.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #pragma hdrstop
 
-void	__stdcall	xrMemFill32_x86		(LPVOID dest, u32 value, size_t count)
+void	__stdcall	xrMemFill32_x86		(LPVOID dest, u32 value, u32 count)
 {
 	u32*	ptr = (u32*) dest;
 	u32*	end = ptr + count;
@@ -9,7 +9,7 @@ void	__stdcall	xrMemFill32_x86		(LPVOID dest, u32 value, size_t count)
 }
 
 #if defined(M_BORLAND) || defined(_M_AMD64)
-void	__stdcall	xrMemFill32_MMX		(LPVOID dest,  u32 value, size_t count)
+void	__stdcall	xrMemFill32_MMX		(LPVOID dest,  u32 value, u32 count)
 {
 	u32*	ptr = (u32*) dest;
 	u32*	end = ptr + count;
